--- AddOns/Intenz_Configs/CfgVehicles/Armband_Base/Cfg.hpp
@@ -1,20 +1,0 @@
-class CfgVehicles
-{
-	class Armband_ColorBase;
-
-	class Intenz_Armband: Armband_ColorBase
-	{
-		displayName="IntenZ Armband";
-		descriptionShort="Armband dedicated to the IntenZ community";
-
-		scope=2;
-		hiddenSelectionsTextures[]=
-		{
-			"Intenz_Images\Textures\IntenzArmband_g_co.paa",
-			"Intenz_Images\Textures\IntenzArmband_co.paa",
-			"Intenz_Images\Textures\IntenzArmband_co.paa",
-			"Intenz_Images\Textures\IntenzArmband_co.paa",
-			"Intenz_Images\Textures\IntenzArmband_co.paa"
-		};
-	};
-};
