+++ AddOns/Intenz_Configs/CfgVehicles/Armbands/Cfg.hpp
@@ -1,0 +1,15 @@
+class Intenz_Armband: Armband_ColorBase
+{
+	displayName="IntenZ Armband";
+	descriptionShort="Armband dedicated to the IntenZ community";
+
+	scope=2;
+	hiddenSelectionsTextures[]=
+	{
+		"\Intenz_Images\Textures\Armbands\IntenzArmband_g_co.paa",
+		"\Intenz_Images\Textures\Armbands\IntenzArmband_co.paa",
+		"\Intenz_Images\Textures\Armbands\IntenzArmband_co.paa",
+		"\Intenz_Images\Textures\Armbands\IntenzArmband_co.paa",
+		"\Intenz_Images\Textures\Armbands\IntenzArmband_co.paa"
+	};
+};
