+++ AddOns/Intenz_Configs/CfgVehicles/Cfg.hpp
@@ -9,6 +9,6 @@ class CfgVehicles
 	class Car;
 	#include "Car\Cfg.hpp"
 	
-	class Armband_Base;
-	#include "Armband_Base\Cfg.hpp"
+	class Armband_ColorBase;
+	#include "Armbands\Cfg.hpp"
 };
